#45  Move create table to DDL package
