+++ cql-parser/src/main/scala/troy/cql/parser/CqlParser.scala
@@ -20,14 +20,14 @@ import troy.cql.ast.dml._
 import troy.cql.ast.dml.{ UpdateParam, UpdateParamValue, UpdateVariable }
 import troy.cql.parser.{ Helpers, TermParser }
 import troy.cql.parser.dml.{ DeleteStatementParser, InsertStatementParser, SelectStatementParser }
-import troy.cql.parser.ddl.CreateKeyspaceParser
+import troy.cql.parser.ddl.{ CreateKeyspaceParser, CreateTableParser }
 import troy.cql.ast.ddl.KeyspaceName
 import scala.util.parsing.combinator._
 
 // Based on CQLv3.4.3: https://cassandra.apache.org/doc/latest/cql/index.html
 object CqlParser extends JavaTokenParsers
     with Helpers with TermParser
-    with CreateKeyspaceParser
+    with CreateKeyspaceParser with CreateTableParser
     with SelectStatementParser with InsertStatementParser with DeleteStatementParser {
   def parseSchema(input: String): ParseResult[Seq[DataDefinition]] =
     parse(phrase(rep(dataDefinition <~ semicolon)), input)
@@ -45,33 +45,6 @@ object CqlParser extends JavaTokenParsers
   def alterKeyspace: Parser[Cql3Statement] = ??? // <create-keyspace-stmt> ::= ALTER KEYSPACE <identifier> WITH <properties>
 
   def dropKeyspace: Parser[Cql3Statement] = ??? // <drop-keyspace-stmt> ::= DROP KEYSPACE ( IF EXISTS )? <identifier>
-
-  def createTable: Parser[CreateTable] = {
-    import Table._
-
-    def createTable = "create".i ~> ("table".i | "columnfamily".i)
-
-    def columnDefinition: Parser[Column] = identifier ~ dataType ~ "STATIC".flag ~ "PRIMARY KEY".flag ^^^^ Column
-
-    def primaryKeyDefinition: Parser[PrimaryKey] = {
-      def simplePartitionKey = identifier.asSeq
-      def compositePartitionKey = parenthesis(rep1sep(identifier, ","))
-      def partitionKeys: Parser[Seq[String]] = simplePartitionKey | compositePartitionKey
-      def clusteringColumns: Parser[Seq[String]] = ("," ~> rep1sep(identifier, ",")) orEmpty
-
-      "PRIMARY KEY".i ~> parenthesis(partitionKeys ~ clusteringColumns) ^^^^ PrimaryKey
-    }
-
-    def option: Parser[CreateTableOption] = ??? // <property> | COMPACT STORAGE | CLUSTERING ORDER
-    def withOptions: Parser[Seq[CreateTableOption]] = ("WITH".i ~> rep1sep(option, "AND".i)) orEmpty
-
-    createTable ~>
-      ifNotExists ~
-      tableName ~
-      ("(" ~> rep1sep(columnDefinition, ",")) ~
-      ("," ~> primaryKeyDefinition).? ~
-      (")" ~> withOptions) ^^^^ CreateTable.apply
-  }
 
   def createIndex: Parser[CreateIndex] = {
     import Index._
