+++ cql-parser/src/test/scala/troy/cql/parser/ddl/CreateTableParserTest.scala
@@ -1,0 +1,29 @@
+package troy.cql.parser.ddl
+
+import org.scalatest._
+import troy.cql.ast.{ CqlParser, CreateTable, DataType }
+
+class CreateTableParserTest extends FlatSpec with Matchers {
+  "Create Table Parser" should "parse simple create table" in {
+    val statement = parseSchemaAs[CreateTable]("CREATE TABLE test.posts (author_id text PRIMARY KEY);")
+    statement.ifNotExists shouldBe false
+    statement.tableName.keyspace.get.name shouldBe "test"
+    statement.tableName.table shouldBe "posts"
+    statement.columns.size shouldBe 1
+    statement.columns.head.dataType shouldBe DataType.text
+    statement.columns.head.isPrimaryKey shouldBe true
+    statement.columns.head.isStatic shouldBe false
+    statement.columns.head.name shouldBe "author_id"
+  }
+
+  def parseSchema(statement: String) =
+    CqlParser.parseSchema(statement) match {
+      case CqlParser.Success(res, _)    => res
+      case CqlParser.Failure(msg, next) => fail(s"Parse Failure: $msg, line = ${next.pos.line}, column = ${next.pos.column}")
+    }
+
+  def parseSchemaAs[T](statement: String) =
+    parseSchema(statement)
+      .head
+      .asInstanceOf[T]
+}
